move the kaplanmeier fitter to the convoys.single
